fix: Netlify 배포 시 새로고침 크래시 방지 - 동적 세그먼트 라우트를 static + query string으로 변경
