+++ apps/admin/src/pages/activity/index.tsx
@@ -1,11 +1,7 @@
 import React, { useEffect } from 'react';
 import { useRouter } from 'next/router';
 import { Flex, Button, Space, TextField, Text } from '@ceos-fe/ui';
-import {
-  RewardResponse,
-  SponsoredByDTO,
-  adminSponsoredByApi,
-} from '@ceos-fe/utils';
+import { RewardResponse, adminActivityApi } from '@ceos-fe/utils';
 import styled from '@emotion/styled';
 import { SponsoredByContainer } from '../../components/sponsoredBy/SponsoredByContainer/index';
 import useInfiniteQueries from '../../hooks/useInfiniteQueries';
@@ -13,52 +9,55 @@ import { PageTitle } from '../../components/Common/PageTitle';
 import { useForm } from 'react-hook-form';
 import { ImageUploader } from '../../components/ImageUploader/index';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { ActivityCardContainer } from '../../components/activity/ActivityCardContainer';
+import {
+  ActivityDTO,
+  ActivityResponse,
+} from '../../../../../packages/utils/src/apis/admin/adminActivityApi';
 
-export default function SponsoredBy() {
+export default function Activity() {
   const queryClient = useQueryClient();
   const router = useRouter();
   const {
     watch,
     register,
     setValue,
-    handleSubmit,
     getValues,
+    handleSubmit,
     formState: { isValid },
   } = useForm();
-  const { infiniteData, ref } = useInfiniteQueries<RewardResponse>({
-    queryKey: ['sponsor'],
+  const { infiniteData, ref } = useInfiniteQueries<ActivityResponse>({
+    queryKey: ['activity'],
     queryFunction: ({ pageParam = 0 }) =>
-      adminSponsoredByApi.GET_SPONSOR({ pageNum: pageParam, limit: 12 }),
-    PageItem: SponsoredByContainer,
+      adminActivityApi.GET_ACTIVITY({ pageNum: pageParam, limit: 12 }),
+    PageItem: ActivityCardContainer,
   });
 
-  // 스폰서 생성 api
-  const postSponsorCreateMutation = useMutation(
-    adminSponsoredByApi.POST_SPONSOR,
+  // 활동 생성 api
+  const postActivityCreateMutation = useMutation(
+    adminActivityApi.POST_ACTIVITY,
     {
       onSuccess: () => {
         alert('추가 완료');
-        queryClient.invalidateQueries([['sponsor']]);
+        queryClient.invalidateQueries([['activity']]);
       },
     },
   );
 
-  // 스폰서 수정 api
-  const patchSponsorCreateMutation = useMutation(
-    adminSponsoredByApi.PAPTCH_SPONSOR,
-    {
-      onSuccess: () => {
-        alert('수정 완료');
-        queryClient.invalidateQueries([['sponsor']]);
-        router.push('/sponsoredby');
-      },
+  // 활동 수정 api
+  const putActivityCreateMutation = useMutation(adminActivityApi.PUT_ACTIVITY, {
+    onSuccess: () => {
+      alert('수정 완료');
+      queryClient.invalidateQueries([['activity']]);
+      router.replace('/activity');
     },
-  );
+  });
 
   // 수정 시 정보 가져오기 api
-  const getSponsorMutation = useMutation(adminSponsoredByApi.GET_ONE_SPONSOR, {
-    onSuccess: async (data: SponsoredByDTO) => {
+  const getActivityMutation = useMutation(adminActivityApi.GET_ONE_ACTIVITY, {
+    onSuccess: async (data: ActivityDTO) => {
       setValue('name', data.name);
+      setValue('content', data.content);
       setValue('imageUrl', data.imageUrl);
     },
     onError: (error: any) => {
@@ -67,15 +66,15 @@ export default function SponsoredBy() {
   });
 
   const onSubmit = (data: any) => {
-    if (Number(router.query[''])) {
-      patchSponsorCreateMutation.mutate({
-        id: Number(Number(router.query[''])),
+    if (Number(router.query.id)) {
+      putActivityCreateMutation.mutate({
+        id: Number(router.query.id),
         payload: {
           ...data,
         },
       });
     } else {
-      postSponsorCreateMutation.mutate({
+      postActivityCreateMutation.mutate({
         payload: {
           ...data,
         },
@@ -84,20 +83,22 @@ export default function SponsoredBy() {
   };
 
   useEffect(() => {
-    if (Number(router.query[''])) {
-      getSponsorMutation.mutate(Number(router.query['']));
+    if (Number(router.query.id)) {
+      getActivityMutation.mutate(Number(router.query.id));
     }
-  }, [router]);
+  }, [router.query.id]);
 
   return (
     <Flex direction="column" align="flex-start" justify="flex-start">
       <Flex align="flex-end" justify="space-between" width={1032}>
         <PageTitle
-          title={'SPONSORED BY'}
-          description={'세오스 후원 단체의 목록을 관리합니다.'}
+          title={'ACTIVITY'}
+          description={'세오스 활동 목록을 관리합니다.'}
         />
       </Flex>
-      <Space height={48} />
+      <div>
+        <Space height={48} />
+      </div>
       <Flex
         direction="row"
         align="flex-start"
@@ -134,10 +135,21 @@ export default function SponsoredBy() {
         </div>
         <div>
           <TextField
-            label="후원 단체명"
+            label="활동 제목"
             placeholder="내용을 입력하세요."
             isAdmin
             {...register('name', {
+              required: true,
+            })}
+          />
+          <Space height={24} />
+          <TextField
+            label="활동 설명"
+            placeholder="내용을 입력하세요."
+            height={64}
+            multiline
+            isAdmin
+            {...register('content', {
               required: true,
             })}
           />
@@ -146,7 +158,7 @@ export default function SponsoredBy() {
             <Text webTypo="Label3">썸네일 이미지</Text>
             <Space height={8} />
             <ImageUploader
-              imageApiType="SPONSOR"
+              imageApiType="ACTIVITY"
               label="imageUrl"
               value={watch('imageUrl')}
               setValue={setValue}
